Add authorization checks and clean up models on delete

+++ src/Http/Controllers/Api/EmailAddressController.php
@@ -3,14 +3,18 @@
 namespace BristolSU\Mail\Http\Controllers\Api;
 
 use BristolSU\Mail\Models\EmailAddress;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 
 class EmailAddressController extends Controller
 {
+    use AuthorizesRequests;
 
     public function store(Request $request)
     {
+        $this->authorize('manage-mail');
+
         $request->validate([
             'email' => 'email:rfc,dns|unique:portal_mail_email_addresses,email'
         ]);
@@ -22,6 +26,8 @@ class EmailAddressController extends Controller
 
     public function destroy(EmailAddress $emailAddress)
     {
+        $this->authorize('manage-mail');
+
         $emailAddress->delete();
 
         return response()->json($emailAddress);
